Deprecate the (semi)old flow control nodes
